requirements.txt
annotated-types==0.6.0
anyio==3.7.1
apt-clone==0.2.1
apturl==0.5.2
attrs==19.3.0
Automat==0.8.0
beautifulsoup4==4.8.2
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
colorama==0.4.6
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
Django==2.2.12
entrypoints==0.3
exceptiongroup==1.1.3
fastapi==0.104.1
greenlet==3.0.1
grpcio==1.16.1
h11==0.14.0
httplib2==0.14.0
hyperlink==19.0.0
idna==3.4
ifaddr==0.1.6
IMDbPY==6.8
incremental==16.10.1
joblib==1.3.2
kazam==1.4.5
keyring==18.0.1
kiwisolver==1.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
mangum==0.17.0
MarkupSafe==1.1.0
matplotlib==3.1.2
nemo-emblems==4.8.1
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
numpy==1.24.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pandas==2.0.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
psycopg2-binary==2.9.7
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
pydantic==2.4.2
pydantic-core==2.10.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==5.0.1
pyOpenSSL==19.0.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-xapp==2.0.2
python-xlib==0.23
pytz==2023.3.post1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
scikit-learn==1.3.2
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
setproctitle==1.1.10
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
soupsieve==1.9.5
SQLAlchemy==2.0.23
sqlparse==0.2.4
starlette==0.27.0
systemd-python==234
threadpoolctl==3.2.0
tinycss2==1.0.2
tldextract==2.2.1
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==27.4
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
uvicorn==0.24.0.post1
vboxapi==1.0
wadllib==1.3.3
webencodings==0.5.1
websockets==8.1
wsproto==0.15.0
xkit==0.0.0
youtube-dl==2021.4.26
zeroconf==0.24.4
zope.interface==4.7.1
